return amounts as a string encoded decimal, move format_lbrycrd_keys function to commands.py
