+++ lib/tests/test_util.py
@@ -1,29 +1,8 @@
 import unittest
-from lib.util import format_satoshis, parse_URI, format_lbrycrd_keys
+from lib.util import format_satoshis, parse_URI
 
 
 class TestUtil(unittest.TestCase):
-    def test_format_lbrycrd_keys(self):
-        a = {'test': 1,
-         'nOut': 1}
-        out = format_lbrycrd_keys(a)
-        self.assertTrue('nout' in out)
-        self.assertFalse('nOut' in out)
-        self.assertEqual(1, out['nout'])
-        self.assertEqual(1, out['test'])
-        a = [{'test': 1,
-          'nOut': 1}, {'test': 1,
-          'nOut': [{'nOut': 1}]}]
-        out = format_lbrycrd_keys(a)
-        self.assertTrue('nout' in out[0])
-        self.assertFalse('nOut' in out[0])
-        self.assertEqual(1, out[0]['nout'])
-        self.assertTrue('nout' in out[1])
-        self.assertFalse('nOut' in out[1])
-        self.assertTrue('nout' in out[1]['nout'][0])
-        self.assertFalse('nOut' in out[1]['nout'][0])
-        self.assertEqual(1, out[1]['nout'][0]['nout'])
-
     def test_format_satoshis(self):
         result = format_satoshis(1234)
         expected = "0.00001234"
